Macro to enable/disable a specific patch.
This commit adds a new file which contains macros which can be used to
enable or disable a patch. Whenever a new patch is added, it will be a
good practice to add macro related to it and surround corresponding code
changes by `#if' and `#endif'. This will allow to disable a patch by
toggling a single bit.